requirements.txt
flask
flask-sqlalchemy
flask-debugtoolbar
mysqlclient
flask_script
requests
uwsgi
qiniu
flask-qiniustorage